Fix timezone oddness in date
We really want to be simply adding the Sydney timezone to the datetime,
not taking the datetime and converting it to the Sydney timezone.

When run locally (in Sydney), these seem equivalent. When you run on
Heroku, the servers are in UTC. Converting to Sydney time adds 11 hours
(for now...) to the times that end up in the database.